requirements.txt
pandas
numpy
matplotlib
plotly
scikit-learn
seaborn
xgboost
flask
cassandra-driver
-e .